:recycle: Refactored the 1k line file for place...
